requirements.txt
langchain
langchain_groq
langserve
fastapi
uvicorn
python-dotenv
pydantic